multauth/api/permissions.py
from rest_framework import permissions


# EXAMPLE
# class IsAdmin(permissions.BasePermission):
#     def has_permission(self, request, view):
#         return request.user.is_authenticated and request.user.is_admin


# EXAMPLE
# class IsCustomUser(permissions.BasePermission):
#     def has_permission(self, request, view):
#         return request.user.is_authenticated and request.user.is_custom_user
